Revert "feat(compiler): bring local wing files (#252)"
This reverts commit 05fb820b3e5fb25b8906527c1501bf12beda0c34.
