--- examples/invalid/circular-bring.w
@@ -1,1 +1,0 @@
-bring "./circular/mod.w";
