--- examples/invalid/circular/mod.w
@@ -1,1 +1,0 @@
-bring "../circular-bring.w";
