--- examples/simple/bring.w
@@ -1,2 +1,0 @@
-bring "./modules/mod1.w";
-bring "./modules/mod2.w";
