--- examples/simple/modules/mod1.w
@@ -1,2 +1,0 @@
-let x = 1;
-let y = 2;
