--- examples/simple/modules/mod2.w
@@ -1,4 +1,0 @@
-bring "./sub/mod.w";
-
-let z = 3;
-let q = 4;
