--- examples/simple/modules/sub/mod.w
@@ -1,1 +1,0 @@
-let w = "what";
