+++ libs/tree-sitter-wing/test/corpus/statements.txt
@@ -254,18 +254,3 @@ struct Test {
         )
     )
 )
-
-==================================
-Bring statement
-==================================
-
-bring cloud;
-bring "./mod.w";
-
----
-(source
-    (short_import_statement
-        module_name: (identifier))
-    (short_import_statement
-        module_name: (string))
-)
