+++ libs/wingc/src/ast.rs
@@ -79,14 +79,9 @@ pub struct Constructor {
 
 #[derive(Debug)]
 pub enum Statement {
-	// TODO: merge these statements into a single one
 	Use {
 		module_name: Symbol, // Reference?
 		identifier: Option<Symbol>,
-	},
-	Bring {
-		module_path: String,
-		statements: Scope,
 	},
 	VariableDef {
 		var_name: Symbol,
