+++ libs/wingc/src/parser.rs
@@ -1,9 +1,5 @@
-pub mod bring;
-
-use bring::bring;
 use std::cell::RefCell;
-use std::collections::{HashMap, HashSet};
-use std::path::PathBuf;
+use std::collections::HashMap;
 use std::{str, vec};
 use tree_sitter::Node;
 
@@ -18,7 +14,6 @@ pub struct Parser<'a> {
 	pub source: &'a [u8],
 	pub source_name: String,
 	pub diagnostics: RefCell<Diagnostics>,
-	pub imports: RefCell<&'a mut HashSet<String>>,
 }
 
 impl Parser<'_> {
@@ -114,39 +109,14 @@ impl Parser<'_> {
 
 	fn build_statement(&self, statement_node: &Node) -> DiagnosticResult<Statement> {
 		match statement_node.kind() {
-			"short_import_statement" => {
-				let module_name_node = statement_node.child_by_field_name("module_name").unwrap();
-				if module_name_node.kind() == "identifier" {
-					Ok(Statement::Use {
-						module_name: self.node_symbol(&module_name_node)?,
-						identifier: if let Some(identifier) = statement_node.child_by_field_name("alias") {
-							Some(self.node_symbol(&identifier)?)
-						} else {
-							None
-						},
-					})
-				} else if module_name_node.kind() == "string" {
-					let bring_target = self.node_text(&module_name_node);
-					let bring_target = &bring_target[1..bring_target.len() - 1];
-					let context_dir = PathBuf::from(&self.source_name);
-					let context_dir = context_dir.parent().unwrap().to_str().unwrap();
-					let brought = bring(bring_target, Some(context_dir), *self.imports.borrow_mut());
-					Ok(Statement::Bring {
-						module_path: bring_target.to_string(),
-						statements: brought
-							.unwrap_or((
-								Scope {
-									statements: vec![],
-									env: None,
-								},
-								Diagnostics::new(),
-							))
-							.0,
-					})
+			"short_import_statement" => Ok(Statement::Use {
+				module_name: self.node_symbol(&statement_node.child_by_field_name("module_name").unwrap())?,
+				identifier: if let Some(identifier) = statement_node.child_by_field_name("alias") {
+					Some(self.node_symbol(&identifier)?)
 				} else {
-					panic!("Unexpected bring type {}", module_name_node.kind())
-				}
-			}
+					None
+				},
+			}),
 			"variable_definition_statement" => {
 				let type_ = if let Some(type_node) = statement_node.child_by_field_name("type") {
 					Some(self.build_type(&type_node)?)
